fix: remove redundancy
