+++ src/ai/backend/storage/context.py
@@ -14,6 +14,7 @@ from ai.backend.common.etcd import AsyncEtcd
 from .abc import AbstractVolume
 from .exception import InvalidVolumeError
 from .vfs import BaseVolume
+from .xfs import XfsVolume
 from .purestorage import FlashBladeVolume
 from .types import VolumeInfo
 
@@ -21,6 +22,7 @@ from .types import VolumeInfo
 BACKENDS: Mapping[str, Type[AbstractVolume]] = {
     'purestorage': FlashBladeVolume,
     'vfs': BaseVolume,
+    'xfs': XfsVolume
 }
