+++ src/ai/backend/storage/vfs/__init__.py
@@ -28,7 +28,21 @@ from ..types import (
     SENTINEL,
     Sentinel,
 )
+from ..exception import (
+    ExecutionError,
+    InvalidAPIParameters
+)
 from ..utils import fstime2datetime
+from ai.backend.common.types import BinarySize
+
+
+async def run(cmd: str) -> str:
+    proc = await asyncio.create_subprocess_shell(
+        cmd, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE)
+    out, err = await proc.communicate()
+    if err:
+        raise ExecutionError(err.decode())
+    return out.decode()
 
 
 class BaseVolume(AbstractVolume):
@@ -122,6 +136,12 @@ class BaseVolume(AbstractVolume):
         await loop.run_in_executor(None, _calc_usage, target_path)
         return VFolderUsage(file_count=total_count, used_bytes=total_size)
 
+    async def get_used_bytes(self, vfid: UUID) -> BinarySize:
+        vfpath = self.mangle_vfpath(vfid)
+        info = await run(f'du -hs {vfpath}')
+        used_bytes, _ = info.split()
+        return BinarySize.from_str(used_bytes)
+
     # ------ vfolder internal operations -------
 
     def scandir(self, vfid: UUID, relpath: PurePosixPath) -> AsyncIterator[DirEntry]:
@@ -193,12 +213,21 @@ class BaseVolume(AbstractVolume):
 
     async def move_file(self, vfid: UUID, src: PurePosixPath, dst: PurePosixPath) -> None:
         src_path = self.sanitize_vfpath(vfid, src)
+        if not src_path.is_file():
+            raise InvalidAPIParameters(msg=f'source path {str(src_path)} is not a file')
         dst_path = self.sanitize_vfpath(vfid, dst)
         loop = asyncio.get_running_loop()
+        await loop.run_in_executor(None, lambda: dst_path.parent.mkdir(parents=True, exist_ok=True))
         await loop.run_in_executor(None, src_path.rename, dst_path)
 
     async def copy_file(self, vfid: UUID, src: PurePosixPath, dst: PurePosixPath) -> None:
-        raise NotImplementedError
+        src_path = self.sanitize_vfpath(vfid, src)
+        if not src_path.is_file():
+            raise InvalidAPIParameters(msg=f'source path {str(src_path)} is not a file')
+        dst_path = self.sanitize_vfpath(vfid, dst)
+        loop = asyncio.get_running_loop()
+        await loop.run_in_executor(None, lambda: dst_path.parent.mkdir(parents=True, exist_ok=True))
+        await loop.run_in_executor(None, lambda: shutil.copyfile(str(src_path), str(dst_path)))
 
     async def prepare_upload(self, vfid: UUID) -> str:
         vfpath = self.mangle_vfpath(vfid)
